// src/types/enums.ts
export enum TaskType {
    ConnectPairsTextText = "Connect_Pairs_Text-Text",
    ConnectPairsTextImage = "Connect_Pairs_Text-Image",
    FourChoicesImage = "Four_Choices_Image-Texts",
    FourChoicesText = "Four_Choices_Text-Images",
}

export enum Difficulties {
    Easy = "Easy",
    Hard = "Hard"
}